langue/requirements.txt
setuptools>=69.0.0
wheel

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1

# LangChain + Gemini
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.6
google-generativeai>=0.7.2

# VectorStore
chromadb>=0.5.0

# Dataset HuggingFace
datasets>=2.19.0
huggingface-hub>=0.23.0

# STT / TTS
openai-whisper>=20231117
elevenlabs>=1.2.0

# FastAPI utils
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.7.0